+++ test_autoarray/structures/triangles/conftest.py
@@ -1,4 +1,19 @@
 from autoarray.numpy_wrapper import numpy as np
 import pytest
 
-from autoarray.structures.triangles.array import ArrayTriangles
+
+@pytest.fixture
+def compare_with_nans():
+    def compare_with_nans_(arr1, arr2):
+        nan_mask1 = np.isnan(arr1)
+        nan_mask2 = np.isnan(arr2)
+
+        equal_elements = np.where(
+            nan_mask1 & nan_mask2,
+            True,
+            arr1 == arr2,
+        )
+
+        return np.all(equal_elements)
+
+    return compare_with_nans_
